rename user-settings folder
