+++ src/web/Routes.jsx
@@ -92,9 +92,9 @@ import TicketDetailsPage from 'web/pages/tickets/DetailsPage';
 import TicketsPage from 'web/pages/tickets/ListPage';
 import TlsCertificateDetailsPage from 'web/pages/tlscertificates/DetailsPage';
 import TlsCertificatesPage from 'web/pages/tlscertificates/ListPage';
+import UserSettingsPage from 'web/pages/user-settings/UserSettingsPage';
 import UserDetailsPage from 'web/pages/users/DetailsPage';
 import UsersPage from 'web/pages/users/ListPage';
-import UserSettingsPage from 'web/pages/usersettings/UserSettingsPage';
 import VulnerabilitiesPage from 'web/pages/vulns/ListPage';
 import {isLoggedIn as selectIsLoggedIn} from 'web/store/usersettings/selectors';
